# chore: refactor reply.send and prioritize kReplyIsError