--- a/lib/error-handler.js
+++ b/lib/error-handler.js
@@ -39,7 +39,7 @@ function handleError (reply, error, cb) {
         if (!reply.log[kDisableRequestLogging]) {
           reply.log.warn(
             { req: reply.request, res: reply, err: error },
-            error && error.message
+            error?.message
           )
         }
         reply.raw.writeHead(reply.raw.statusCode)
@@ -89,14 +89,14 @@ function defaultErrorHandler (error, request, reply) {
     if (!reply.log[kDisableRequestLogging]) {
       reply.log.info(
         { res: reply, err: error },
-        error && error.message
+        error?.message
       )
     }
   } else {
     if (!reply.log[kDisableRequestLogging]) {
       reply.log.error(
         { req: request, res: reply, err: error },
-        error && error.message
+        error?.message
       )
     }
   }