--- a/lib/reply.js
+++ b/lib/reply.js
@@ -126,16 +126,16 @@ Reply.prototype.hijack = function () {
 }
 
 Reply.prototype.send = function (payload) {
-  if (this[kReplyIsRunningOnErrorHook] === true) {
+  if (this[kReplyIsRunningOnErrorHook]) {
     throw new FST_ERR_SEND_INSIDE_ONERR()
   }
 
-  if (this.sent) {
+  if (this.sent === true) {
     this.log.warn({ err: new FST_ERR_REP_ALREADY_SENT(this.request.url, this.request.method) })
     return this
   }
 
-  if (payload instanceof Error || this[kReplyIsError] === true) {
+  if (this[kReplyIsError] || payload instanceof Error) {
     this[kReplyIsError] = false
     onErrorHook(this, payload, onSendHook)
     return this
@@ -162,8 +162,8 @@ Reply.prototype.send = function (payload) {
       return this
     }
 
-    if (payload?.buffer instanceof ArrayBuffer) {
-      if (hasContentType === false) {
+    if (payload.buffer instanceof ArrayBuffer) {
+      if (!hasContentType) {
         this[kReplyHeaders]['content-type'] = CONTENT_TYPE.OCTET
       }
       const payloadToSend = Buffer.isBuffer(payload) ? payload : Buffer.from(payload.buffer, payload.byteOffset, payload.byteLength)
@@ -171,7 +171,7 @@ Reply.prototype.send = function (payload) {
       return this
     }
 
-    if (hasContentType === false && typeof payload === 'string') {
+    if (!hasContentType && typeof payload === 'string') {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.PLAIN
       onSendHook(this, payload)
       return this
@@ -182,26 +182,24 @@ Reply.prototype.send = function (payload) {
     if (typeof payload !== 'string') {
       preSerializationHook(this, payload)
       return this
-    } else {
-      payload = this[kReplySerializer](payload)
     }
+    payload = this[kReplySerializer](payload)
 
     // The indexOf below also matches custom json mimetypes such as 'application/hal+json' or 'application/ld+json'
-  } else if (hasContentType === false || contentType.indexOf('json') > -1) {
-    if (hasContentType === false) {
+  } else if (!hasContentType || contentType.indexOf('json') !== -1) {
+    if (!hasContentType) {
       this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
-    } else {
+    } else if (contentType.indexOf('charset') === -1) {
       // If user doesn't set charset, we will set charset to utf-8
-      if (contentType.indexOf('charset') === -1) {
-        const customContentType = contentType.trim()
-        if (customContentType.endsWith(';')) {
-          // custom content-type is ended with ';'
-          this[kReplyHeaders]['content-type'] = `${customContentType} charset=utf-8`
-        } else {
-          this[kReplyHeaders]['content-type'] = `${customContentType}; charset=utf-8`
-        }
+      const customContentType = contentType.trim()
+      if (customContentType.endsWith(';')) {
+        // custom content-type is ended with ';'
+        this[kReplyHeaders]['content-type'] = `${customContentType} charset=utf-8`
+      } else {
+        this[kReplyHeaders]['content-type'] = `${customContentType}; charset=utf-8`
       }
     }
+
     if (typeof payload !== 'string') {
       preSerializationHook(this, payload)
       return this