- Refactors `reply.send` to prioritize `this[kReplyIsError]`, which is a simple boolean check, over `instanceof Error`

https://github.com/fastify/fastify/blob/afd7f509daec4d43ef9dac9d3ea1c8a0cf1ea092/lib/reply.js#L138

This should run faster when we throw one of our errors, and if it's not the case, the boolean check should be pretty much instant so it's another small win

- Also simplifies the branching of the method

New benchmark, `benchmark:parser:error`:

PR:
```sh
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] 
[1] ┌─────────┬────────┬──────────┬──────────┬──────────┬─────────────┬────────────┬──────────┐
[1] │ Stat    │ 2.5%   │ 50%      │ 97.5%    │ 99%      │ Avg         │ Stdev      │ Max      │
[1] ├─────────┼────────┼──────────┼──────────┼──────────┼─────────────┼────────────┼──────────┤
[1] │ Latency │ 584 ms │ 10221 ms │ 26984 ms │ 27559 ms │ 11858.67 ms │ 8067.32 ms │ 28066 ms │
[1] └─────────┴────────┴──────────┴──────────┴──────────┴─────────────┴────────────┴──────────┘
[1] ┌───────────┬────────┬────────┬─────────┬─────────┬───────────┬──────────┬────────┐
[1] │ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg       │ Stdev    │ Min    │
[1] ├───────────┼────────┼────────┼─────────┼─────────┼───────────┼──────────┼────────┤
[1] │ Req/Sec   │ 6,871  │ 6,871  │ 8,415   │ 14,255  │ 10,026.94 │ 2,702.45 │ 6,868  │
[1] ├───────────┼────────┼────────┼─────────┼─────────┼───────────┼──────────┼────────┤
[1] │ Bytes/Sec │ 996 kB │ 996 kB │ 1.22 MB │ 2.07 MB │ 1.45 MB   │ 392 kB   │ 996 kB │
[1] └───────────┴────────┴────────┴─────────┴─────────┴───────────┴──────────┴────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1] 
[1] 0 2xx responses, 300798 non 2xx responses
[1] 3309k requests in 30.06s, 43.6 MB read
```
main:

```sh
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[1] 
[1] ┌─────────┬────────┬──────────┬──────────┬──────────┬─────────────┬────────────┬──────────┐
[1] │ Stat    │ 2.5%   │ 50%      │ 97.5%    │ 99%      │ Avg         │ Stdev      │ Max      │
[1] ├─────────┼────────┼──────────┼──────────┼──────────┼─────────────┼────────────┼──────────┤
[1] │ Latency │ 562 ms │ 10204 ms │ 26899 ms │ 27451 ms │ 11830.13 ms │ 7992.12 ms │ 27964 ms │
[1] └─────────┴────────┴──────────┴──────────┴──────────┴─────────────┴────────────┴──────────┘
[1] ┌───────────┬────────┬────────┬─────────┬────────┬──────────┬──────────┬────────┐
[1] │ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%  │ Avg      │ Stdev    │ Min    │
[1] ├───────────┼────────┼────────┼─────────┼────────┼──────────┼──────────┼────────┤
[1] │ Req/Sec   │ 6,539  │ 6,539  │ 7,911   │ 13,807 │ 9,786.87 │ 2,691.72 │ 6,538  │
[1] ├───────────┼────────┼────────┼─────────┼────────┼──────────┼──────────┼────────┤
[1] │ Bytes/Sec │ 948 kB │ 948 kB │ 1.15 MB │ 2 MB   │ 1.42 MB  │ 390 kB   │ 948 kB │
[1] └───────────┴────────┴────────┴─────────┴────────┴──────────┴──────────┴────────┘
[1] 
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1] 
[1] 0 2xx responses, 293605 non 2xx responses
[1] 3230k requests in 30.05s, 42.6 MB read
```